Add resetCmd
