+++ pkg/install/users.go
@@ -5,12 +5,30 @@ import (
 	"os"
 	"os/exec"
 	"os/user"
+	"reflect"
 	"strings"
 
 	"github.com/sirupsen/logrus"
 
 	"github.com/k0sproject/k0s/internal/util"
+	"github.com/k0sproject/k0s/pkg/apis/v1beta1"
+	"github.com/k0sproject/k0s/pkg/constant"
 )
+
+// CreateControllerUsers accepts a cluster config, and cfgVars and creates controller users accordingly
+func CreateControllerUsers(clusterConfig *v1beta1.ClusterConfig, k0sVars constant.CfgVars) error {
+	users := getUserList(*clusterConfig.Install.SystemUsers)
+	var messages []string
+	for _, v := range users {
+		if err := EnsureUser(v, k0sVars.DataDir); err != nil {
+			messages = append(messages, err.Error())
+		}
+	}
+	if len(messages) > 0 {
+		return fmt.Errorf(strings.Join(messages, "\n"))
+	}
+	return nil
+}
 
 // EnsureUser checks if a user exists, and creates it, if it doesn't
 // TODO: we should also consider modifying the user, if the user exists, but with wrong settings
@@ -84,4 +102,15 @@ func quoteCmd(cmd *exec.Cmd) string {
 		q = append(q, fmt.Sprintf("%q", s))
 	}
 	return strings.Join(q, ` `)
+}
+
+// get user list
+func getUserList(sysUsers v1beta1.SystemUser) []string {
+	v := reflect.ValueOf(sysUsers)
+	values := make([]string, v.NumField())
+
+	for i := 0; i < v.NumField(); i++ {
+		values[i] = v.Field(i).String()
+	}
+	return values
 }
